Moved import file parsing into another module
